Code is finished

+++ app.py
@@ -93,56 +93,88 @@ async def find_car_by_ID_and_return(id: int, current_user: Optional[models.User]
 
 # Method that adds new car to the database.
 @app.post("/add_car")
-async def add_car(car: models.Cars):
+async def add_car(car: models.Cars, current_user: Optional[models.User] = Depends(get_current_user)):
     try:
-        await db.add_car(car)
-        return {"message": "Car added successfully"}
+        if current_user:
+            user = await current_user
+            if (user.role == "Admin"):
+                await db.add_car(car)
+                return {"message": "Car added successfully"}
+            else:
+                return {"Message": "You are not authorized to add cars, Only admin can do that!!!"}
     except Exception as e:
         return {"message": "Something went wrong while adding new car", "error": str(e)}
 
 # Method that will delete selected car.   
 @app.delete("/delete_car/{car_id}")
-async def delete_car(car_id: int):
+async def delete_car(car_id: int, current_user: Optional[models.User] = Depends(get_current_user)):
     try:
-        await db.delete_car(car_id)
-        return {"message": "Car deleted successfully"}
+        if current_user:
+            user = await current_user
+            if (user.role == "Admin"):
+                await db.delete_car(car_id)
+                await db.admin_delete_user_car(car_id)
+                return {"message": "Car deleted successfully"}
+            else:
+                return {"Message": "You are not authorized to add cars, Only admin can do that!!!"}
     except Exception as e:
         return {"message": "Something went wrong while deleting a car", "error": str(e)}
     
 # Users.............................................................................................
 # Method that will show us all users we have
 @app.get("/Users")
-async def all_users():
+async def all_users(current_user: Optional[models.User] = Depends(get_current_user)):
     try:
-        users = await db.get_users()  # Add parentheses to call the method
-        return users
+        if current_user:
+            user = await current_user
+            if (user.role == "Admin"):
+                users = await db.get_users()
+                return users
+            else:
+                return {"Message": "You are not authorized to add cars, Only admin can do that!!!"}
     except Exception as e:
         return {"message": "Failed to retrieve users from the database.", "error": str(e)}
 
 # Method that will allow us to create new user
 @app.post("/add_user")
-async def add_user(user: models.User):
+async def add_user(user: models.User, current_user: Optional[models.User] = Depends(get_current_user)):
     try:
-        await db.add_user(user)
-        return {"message": "User added successfully"}
+        if current_user:
+            active_user = await current_user
+            if (active_user.role == "Admin" or active_user.role == "Employee"):
+                await db.add_user(user)
+                return {"message": "User added successfully"}
+            else:
+                return {"Message": "You are not authorized to add cars, Only admin or employee can do that!!!"}
     except Exception as e:
         return {"message": "Something went wrong while adding a new user", "error": str(e)}
 
 
 # Method that will allow us to delete some specific user.
 @app.delete("/delete_user/{user_id}")
-async def delete_user(user_id: int):
+async def delete_user(user_id: int, current_user: Optional[models.User] = Depends(get_current_user)):
     try:
-        await db.delete_user(user_id)
-        return {"message": "User deleted successfully"}
+        if current_user:
+            user = await current_user
+            if (user.role == "Admin" or user.role == "Employee"):
+                await db.delete_user(user_id)
+                await db.admin_delete_user_car_through_user(user_id)
+                return {"message": "User deleted successfully"}
+            else:
+                return {"Message": "You are not authorized to add cars, Only admin or employee can do that!!!"}
     except Exception as e:
         return {"message": "Something went wrong while deleting a User", "error": str(e)}
 
 # Method that gets any user by its email
 @app.get("/get_user_by_mail/{mail}")
-async def get_user_email(user_mail: str):
+async def get_user_email(user_mail: str, current_user: Optional[models.User] = Depends(get_current_user)):
     try:
-        user = await db.get_user_by_email(user_mail)
-        return user
+        if current_user:
+            user = await current_user
+            if (user.role == "Admin"):
+                user = await db.find_curent_user_by_email(user_mail)
+                return user
+            else:
+                return {"Message": "You are not authorized to add cars, Only admin can do that!!!"}
     except Exception as e:
         return {"message": "Failed to retrieve users from the database.", "error": str(e)}
